r/aws_glue_schema: Reduce visibility.
